bot/requirements.txt
discord.py==2.3.2
aio_pika==6.10.0
python-dotenv==1.0.0
requests
riotwatcher==3.3.1
PyNaCl==1.5.0